finished UI design admin page , login, sign-up
