+++ package.json
@@ -5,11 +5,13 @@
   "author": "Mashkour Abdel Aziz",
   "license": "MIT",
   "dependencies": {
-    "bootstrap": "4.1.3",
+    "@material-ui/core": "^4.1.0",
+    "@material-ui/icons": "^4.1.0",
     "i18next": "11.9.1",
     "jquery": "^3.4.1",
-    "react": "16.5.2",
-    "react-dom": "16.5.2",
+    "material-ui-next": "^1.0.0-beta.47",
+    "react": "^16.8.0",
+    "react-dom": "^16.8.0",
     "react-i18next": "8.0.7",
     "react-redux": "^7.0.3",
     "react-router-dom": "^5.0.1",
@@ -19,8 +21,6 @@
     "redux": "^4.0.1",
     "redux-thunk": "^2.3.0",
     "seamless-immutable": "^7.1.4",
-    "semantic-ui-css": "^2.4.1",
-    "semantic-ui-react": "^0.87.1",
     "styled-components": "^4.3.1"
   },
   "scripts": {
