+++ src/index.js
@@ -1,6 +1,5 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import 'bootstrap/dist/css/bootstrap.css';
 import { I18nextProvider } from 'react-i18next';
 import i18n from './config/i18n';
 import App from './App';
